Add files via upload
The task was to perform as a Senior Marketing Analyst at Maven Cafe, who just run a test by sending different combinations of promotional offers to existing rewards members. Now that the 30-day period for the test has concluded, the task is 
to identify key customer segments 
and develop a data-driven strategy for future promotional messaging & targeting.